bpo-13501 Build or disable readline with Editline
I have ported the work done by martin.panter to latest master.

 * Include <editline/readline.h> if Editline is configured
 * Redefine HAVE_LIBREADLINE to also indicate if -ledit is used
 * New flag WITH_EDITLINE to differentiate between -lreadline and -ledit
 * Remove conditional completion_matches() declarations; use the declaration
   from the library rather than trying to guess it

Resolves bpo-13501

+++ setup.py
@@ -830,7 +830,6 @@ class PyBuildExt(build_ext):
 
     def detect_readline_curses(self):
         # readline
-        do_readline = self.compiler.find_library_file(self.lib_dirs, 'readline')
         readline_termcap_library = ""
         curses_library = ""
         # Cannot use os.popen here in py3k.
@@ -838,29 +837,38 @@ class PyBuildExt(build_ext):
         if not os.path.exists(self.build_temp):
             os.makedirs(self.build_temp)
         # Determine if readline is already linked against curses or tinfo.
-        if do_readline:
-            if CROSS_COMPILING:
-                ret = os.system("%s -d %s | grep '(NEEDED)' > %s" \
-                                % (sysconfig.get_config_var('READELF'),
-                                   do_readline, tmpfile))
-            elif find_executable('ldd'):
-                ret = os.system("ldd %s > %s" % (do_readline, tmpfile))
+        if sysconfig.get_config_var('HAVE_LIBREADLINE'):
+            if sysconfig.get_config_var('WITH_EDITLINE'):
+                readline_lib = 'edit'
             else:
-                ret = 256
-            if ret >> 8 == 0:
-                with open(tmpfile) as fp:
-                    for ln in fp:
-                        if 'curses' in ln:
-                            readline_termcap_library = re.sub(
-                                r'.*lib(n?cursesw?)\.so.*', r'\1', ln
-                            ).rstrip()
-                            break
-                        # termcap interface split out from ncurses
-                        if 'tinfo' in ln:
-                            readline_termcap_library = 'tinfo'
-                            break
-            if os.path.exists(tmpfile):
-                os.unlink(tmpfile)
+                readline_lib = 'readline'
+            do_readline = self.compiler.find_library_file(self.lib_dirs,
+                readline_lib)
+            if do_readline:
+                if CROSS_COMPILING:
+                    ret = os.system("%s -d %s | grep '(NEEDED)' > %s" \
+                                    % (sysconfig.get_config_var('READELF'),
+                                       do_readline, tmpfile))
+                elif find_executable('ldd'):
+                    ret = os.system("ldd %s > %s" % (do_readline, tmpfile))
+                else:
+                    ret = 256
+                if ret >> 8 == 0:
+                    with open(tmpfile) as fp:
+                        for ln in fp:
+                            if 'curses' in ln:
+                                readline_termcap_library = re.sub(
+                                    r'.*lib(n?cursesw?)\.so.*', r'\1', ln
+                                ).rstrip()
+                                break
+                            # termcap interface split out from ncurses
+                            if 'tinfo' in ln:
+                                readline_termcap_library = 'tinfo'
+                                break
+                if os.path.exists(tmpfile):
+                    os.unlink(tmpfile)
+        else:
+            do_readline = False
         # Issue 7384: If readline is already linked against curses,
         # use the same library for the readline and curses modules.
         if 'curses' in readline_termcap_library:
@@ -896,7 +904,7 @@ class PyBuildExt(build_ext):
             else:
                 readline_extra_link_args = ()
 
-            readline_libs = ['readline']
+            readline_libs = [readline_lib]
             if readline_termcap_library:
                 pass # Issue 7384: Already linked against curses or tinfo.
             elif curses_library:
